+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add KubeEventsLogger option and related methods in loggedCLuster package.
 - Add `events-logger` flag in the operator.
 - Add toggle for `events-logger` in observability-bundle configmap.
+- Add tests for `alloy-events` in events-logger-config.
 
 ### Changed